api-server/requirements.txt
certifi==2018.4.16
Django==2.0.7
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
olefile==0.45.1
Pillow==5.2.0
psycopg2==2.7.5
PyJWT==1.6.4
pytz==2018.5
wincertstore==0.2